anitrack/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.9
Pillow==12.2.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
requests==2.34.1